adding rarity display and name display

+++ Main.py
@@ -5,7 +5,7 @@ if __name__ == '__main__':
     file = File()
     api_arr = []
     for i in range(file.size):
-        api_arr.append(Api(int(file.price[i]), file.link[i]))
+        api_arr.append(Api(float(file.price[i]), file.link[i]))
     while True:
         for i in range(file.size):
             api_arr[i].retrive_and_check()
